requirements.txt
web3==5.31.1
pymongo==4.3.3
graphene==3.2.1
sentry-sdk==1.14.0
redis==4.5.4
pika==1.3.1
pytest==7.3.1
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.1
prometheus-client==0.12.0
pydantic==1.8.2
python-jose==3.3.0
passlib==1.7.4